feat: 선택한 날짜의 상태 데이터 관리 (#38)
