+++ front/src/pages/Main/index.jsx
@@ -3,13 +3,14 @@ import { Outlet } from 'react-router-dom';
 import styles from './style.module.css';
 import NavBar from '../../components/Navbar';
 import SideBar from '../../components/SideBar';
+import SideNav from '../../components/SideNav';
 
 const Index = () => {
   /* NavBar */
   let [activeClass, setActiveClass] = useState(); //일정 리스트에 클레스 추가
   let [dateActive, setDateActive] = useState(false); //일정 리스트 true ,false 확인
-  let [userClassAdd, setUserClassAdd] = useState(); //일정 리스트에 클레스 추가
-  let [userActive, setUserActive] = useState(false); //일정 리스트 true ,false 확인
+  let [userClassAdd, setUserClassAdd] = useState(); //사용자 팝업창에 class 추가
+  let [userActive, setUserActive] = useState(false); //사용자 팝업창 true ,false 확인
   //body클릭시 일정 리스트 모달창 닫기
   document.body.addEventListener('click', () => {
     setDateActive(false);
@@ -34,6 +35,7 @@ const Index = () => {
         <article className={styles.article}>
           <SideBar />
           <Outlet></Outlet>
+          <SideNav></SideNav>
         </article>
       </section>
     </div>
