+++ front/src/pages/Main/style.module.css
@@ -1,2 +1,3 @@
-.section{width: 100%;height: 100vh;}
-.article{display: flex;}
+.section{width: 100%;height: 100vh; overflow: hidden;}
+.article{display: flex; justify-content: space-between;}
+
